improves quarkus devfile by setting host and logging manager explicility. logging manager was always recommended, and host became required after 1.7+
in addition removed external git starter as it was never going to be updated
fast enough.

instead added a community and product zip download using code.quarkus.io
and code.quarkus.redhat.com respectively. This means devfile will always
pick up the latest default supported version and that get the example
to have openshift, metrics and health default enabled.

we might want to have a kubernates version too - or even leave that extension
out and let it be up to the user to handle.